refactor: Move files for authentication data refactor
