+++ test_runner/src/main/kotlin/ftl/adapter/GooglePerformanceMetricsFetch.kt
@@ -3,7 +3,7 @@ package ftl.adapter
 import ftl.adapter.google.toApiModel
 import ftl.adapter.google.toClientModel
 import ftl.api.PerfMetrics
-import ftl.gc.GcToolResults
+import ftl.client.google.GcToolResults
 
 object GooglePerformanceMetricsFetch :
     PerfMetrics.Fetch,
