+++ test_runner/src/main/kotlin/ftl/api/UserAuthorization.kt
@@ -1,4 +1,4 @@
-package ftl.data
+package ftl.api
 
 import ftl.adapter.GoogleUserAuthorizationRequest
