+++ test_runner/src/main/kotlin/ftl/client/google/GcTestMatrix.kt
@@ -2,7 +2,6 @@ package ftl.client.google
 
 import com.google.testing.model.CancelTestMatrixResponse
 import com.google.testing.model.TestMatrix
-import ftl.gc.GcTesting
 import ftl.http.executeWithRetry
 import ftl.run.exception.FlankGeneralError
 import kotlinx.coroutines.Dispatchers
