+++ test_runner/src/main/kotlin/ftl/client/google/OsVersion.kt
@@ -2,7 +2,6 @@ package ftl.client.google
 
 import com.google.testing.model.AndroidVersion
 import com.google.testing.model.IosVersion
-import ftl.gc.GcTesting
 import ftl.http.executeWithRetry
 
 fun androidOsVersions(projectId: String): List<AndroidVersion> =
