+++ test_runner/src/main/kotlin/ftl/client/google/ProvidedSoftwareCatalog.kt
@@ -1,7 +1,6 @@
 package ftl.client.google
 
 import ftl.environment.common.toCliTable
-import ftl.gc.GcTesting
 import ftl.http.executeWithRetry
 
 fun providedSoftwareAsTable() = getProvidedSoftware().toCliTable()
