+++ test_runner/src/main/kotlin/ftl/gc/GcAndroidTestMatrix.kt
@@ -15,6 +15,7 @@ import com.google.testing.model.ToolResultsHistory
 import flank.common.join
 import ftl.args.AndroidArgs
 import ftl.args.isDontAutograntPermissions
+import ftl.client.google.GcTesting
 import ftl.gc.android.mapGcsPathsToApks
 import ftl.gc.android.mapToDeviceFiles
 import ftl.gc.android.mapToDeviceObbFiles
