+++ test_runner/src/main/kotlin/ftl/gc/GcIosTestMatrix.kt
@@ -11,6 +11,7 @@ import com.google.testing.model.TestMatrix
 import com.google.testing.model.TestSpecification
 import com.google.testing.model.ToolResultsHistory
 import ftl.args.IosArgs
+import ftl.client.google.GcTesting
 import ftl.gc.android.mapGcsPathsToFileReference
 import ftl.gc.android.mapToIosDeviceFiles
 import ftl.gc.android.toIosDeviceFile
