+++ test_runner/src/main/kotlin/ftl/run/platform/common/BeforeRunTests.kt
@@ -3,9 +3,9 @@ package ftl.run.platform.common
 import flank.common.logLn
 import ftl.args.IArgs
 import ftl.client.google.GcStorage
+import ftl.client.google.GcTesting
+import ftl.client.google.GcToolResults
 import ftl.config.FtlConstants
-import ftl.gc.GcTesting
-import ftl.gc.GcToolResults
 import ftl.run.exception.FlankGeneralError
 import ftl.util.StopWatch
 import java.io.File
