+++ test_runner/src/test/kotlin/ftl/reports/api/CreateTestExecutionDataKtTest.kt
@@ -6,9 +6,9 @@ import com.google.api.services.toolresults.model.TestCase
 import com.google.api.services.toolresults.model.Timestamp
 import com.google.testing.model.TestExecution
 import com.google.testing.model.ToolResultsStep
+import ftl.client.google.GcToolResults
 import ftl.client.junit.TestExecutionData
 import ftl.client.junit.createTestExecutionDataListAsync
-import ftl.gc.GcToolResults
 import io.mockk.every
 import io.mockk.mockkObject
 import io.mockk.unmockkObject
